view communication ok by bk. kj

+++ Ariadne/ComPlatform.cpp
@@ -112,13 +112,14 @@ void ComPlatform::setWritePram(BYTE* writeBuffer)
 #define speedKP		1.0
 #define steerRatio  1.0
 
-int ComPlatform::setData_steering(int desired_steering)
+int ComPlatform::setData_steering(double desired_steering)
 {
 	int present_steering = dataContainer->getValue_PtoU_STEER(); //-2000~2000
 	desired_steering *= -71;
 	int dif = desired_steering - present_steering;
 
 	desired_steering = (double)desired_steering*steeringKP;
+
 
 	if (desired_steering > 2000) {
 		desired_steering = 2000;
@@ -130,7 +131,7 @@ int ComPlatform::setData_steering(int desired_steering)
 	//dataContainer->setValue_UtoP_STEER(desired_steering);
 }
 
-int ComPlatform::setData_speed(int desired_speed)
+int ComPlatform::setData_speed(double desired_speed)
 {
 	int present_speed = dataContainer->getValue_PtoU_SPEED(); //0~200
 	double speedratio = dataContainer->getValue_yolo_speed_ratio();
